Port scala-xml spec to MUnit
